Move to the right namespace, remove duplicated attribute.
